[#7] refact : index export setting
