+++ src/hooks/useWsTrade.ts
@@ -1,38 +1,33 @@
-import {ITicker, ImarketCodes} from './interfaces';
+import {ITrade, ImarketCodes} from '../interfaces';
 import {useRef, useState, useCallback, useEffect} from 'react';
-import getLastBuffers from './functions/getLastBuffers';
-import sortBuffers from './functions/sortBuffers';
 import {throttle} from 'lodash';
-import socketDataEncoder from './functions/socketDataEncoder';
-import updateSocketData from './functions/updateSocketData';
 
-export function useWsTicker(
+import updateQueueBuffer from '../functions/updateQueueBuffer';
+import socketDataEncoder from '../functions/socketDataEncoder';
+
+function useWsTrade(
   targetMarketCodes: ImarketCodes[],
-  options = {throttle_time: 400},
+  options = {throttle_time: 400, max_length_queue: 100},
 ) {
   const SOCKET_URL = 'wss://api.upbit.com/websocket/v1';
-  const {throttle_time} = options;
+  const {throttle_time, max_length_queue} = options;
   const socket = useRef<WebSocket | null>(null);
-  const buffer = useRef<ITicker[]>([]);
+  const buffer = useRef<ITrade[]>([]);
 
   const [isConnected, setIsConnected] = useState<boolean>(false);
-  const [loadingBuffer, setLoadingBuffer] = useState<ITicker[]>([]);
-  const [socketData, setSocketData] = useState<ITicker[]>();
+  const [socketData, setSocketData] = useState<ITrade[]>();
 
   const throttled = useCallback(
     throttle(() => {
       try {
-        const lastBuffers = getLastBuffers(
+        const updatedBuffer = updateQueueBuffer(
           buffer.current,
-          targetMarketCodes.length,
+          max_length_queue,
         );
-        const sortedBuffers =
-          lastBuffers && sortBuffers(lastBuffers, targetMarketCodes);
-        sortedBuffers && setLoadingBuffer(sortedBuffers);
-        buffer.current = [];
+        buffer.current = updatedBuffer;
+        setSocketData(updatedBuffer);
       } catch (error) {
-        console.error(error);
-        return;
+        throw new Error();
       }
     }, throttle_time),
     [targetMarketCodes],
@@ -40,18 +35,25 @@ export function useWsTicker(
   // socket 세팅
   useEffect(() => {
     try {
+      if (targetMarketCodes.length > 1) {
+        console.error(
+          "[Error] | 'Length' of Target Market Codes should be only 'one' in 'orderbook' and 'trade'. you can request only 1 marketcode's data, when you want to get 'orderbook' or 'trade' data.",
+        );
+        throw new Error();
+      }
+
       if (targetMarketCodes.length > 0 && !socket.current) {
         socket.current = new WebSocket(SOCKET_URL);
         socket.current.binaryType = 'arraybuffer';
 
         const socketOpenHandler = () => {
           setIsConnected(true);
-          console.log('[연결완료] | socket Open Type: ', 'ticker');
+          console.log('[연결완료] | socket Open Type: ', 'trade');
           if (socket.current?.readyState == 1) {
             const sendContent = [
               {ticket: 'test'},
               {
-                type: 'ticker',
+                type: 'trade',
                 codes: targetMarketCodes.map(code => code.market),
               },
             ];
@@ -62,7 +64,6 @@ export function useWsTicker(
 
         const socketCloseHandler = () => {
           setIsConnected(false);
-          setLoadingBuffer([]);
           setSocketData([]);
           buffer.current = [];
           console.log('연결종료');
@@ -74,11 +75,9 @@ export function useWsTicker(
         };
 
         const socketMessageHandler = (evt: MessageEvent<ArrayBuffer>) => {
-          const data = socketDataEncoder<ITicker>(evt.data);
-          if (data) {
-            buffer.current.push(data);
-            throttled();
-          }
+          const data = socketDataEncoder<ITrade>(evt.data);
+          data && buffer.current.push(data);
+          throttled();
         };
 
         socket.current.onopen = socketOpenHandler;
@@ -99,22 +98,7 @@ export function useWsTicker(
     }
   }, [targetMarketCodes]);
 
-  useEffect(() => {
-    try {
-      if (loadingBuffer.length > 0) {
-        if (!socketData) {
-          setSocketData(loadingBuffer);
-        } else {
-          setSocketData(prev => {
-            return prev && updateSocketData(prev, loadingBuffer);
-          });
-          setLoadingBuffer([]);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  }, [loadingBuffer]);
-
   return {socket: socket.current, isConnected, socketData};
 }
+
+export default useWsTrade;
